Got the mean, standard deviation, variance of aboard/fatalities column and added a scatter plot
Here I found the mean, variance and the SD of the aboard and fatalities
column.  I then found the differences between the means of the two
columns which can become helpful later.  I also used the summary
function to obtain some statistics which can also become helpful later.
And finally, I included a scatter plot to determine the relationship
between the number of passengers aboard and the fatalities.