Analyzed the data to return the survivor to death ratio/percentage
I have found out what is the survivor to death ratio of each accident
along with the percentage of each.  I have found out the average
percentage of accidents where all passengers died, and I also created a
user defined function where you can enter the % and it will return the
number of accidents that occured with that death rate.