Graphed the top 10 locations Douglas DC-3 has crashed at in the world
Since Douglas DC-3 is the plane that crashes the most overall, I wanted
to see if there was some kind of relationship in regards where it
crashed at in the world.  By plotting the top 10 crash locations by the
number of fatalities on the map, hopefully that will tell us if the
locations are just concetrated on just on single region or scatted
across the world.